requirements.txt
scikit-learn==1.5.2
requests==2.32.3
pyyaml==6.0.2
cryptography==43.0.1